requirements.txt
langchain
langchain-community
langchain-qdrant
langchain-huggingface
langchain-ollama
sentence-transformers
qdrant-client
psycopg2-binary
redis
streamlit
